🫧 refactor: custom hook - useLocationTracker(사용자 위치추적), useMission(서버통신) 추가
